Minor fixes to NPMRDS ETL process.

+++ src/data_types/npmrds/dt-npmrds_travel_times_imp/tasks/compute_statistics.ts
@@ -38,10 +38,10 @@ export default async function main(
             SELECT
                 frc,
                 jsonb_build_object(
-                  'total_tmcs',             COUNT( DISTINCT tmc ),
-                  'total_records',          COUNT(1),
+                  'total_tmcs',               COUNT( DISTINCT tmc ),
+                  'total_records',            COUNT(1),
                   'tmcs_md5sum',              MD5(string_agg( DISTINCT tmc, '|' ORDER BY TMC)),
-                  'travel_time_all_sum',      SUM(travel_time_all_vehicles)::TEXT
+                  'travel_time_all_sum',      SUM(travel_time_all_vehicles::NUMERIC)::TEXT
                 ) AS summary
               FROM %I.%I
                 INNER JOIN %I.%I
